requirements.txt
Flask>=2.1
requests
beautifulsoup4
pandas
nltk
reportlab
gspread
google-auth
transformers>=4.0.0
